backend/requirements.txt
django==4.1.7
djangorestframework==3.14.0
django-filter==22.1
django_cleanup==7.0.0
django-adminlte-3==0.1.6
django-telethon==1.1.5
python-dotenv==0.21.1